refactor: 각 api 분리
