+++ src/services/article/apis/constants.ts
@@ -1,0 +1,2 @@
+export const defaultArticlesPageCursor = -1;
+export const defaultArticlesPageSize = 20;
