+++ src/services/article/apis/createArticle.ts
@@ -1,0 +1,40 @@
+import type { ApiSuccessResponse } from '~/types';
+
+import { endpoints } from '~/react-query/common';
+import { privateAxios } from '~/utils/axios';
+
+export interface CreateArticleImagePayload {
+  imagePath: string; // 이미지를 삭제하기 위한 URL
+  imageUrl: string; // 이미지를 렌더링하기 위한 URL
+}
+
+export interface CreateArticleParams {
+  categoryId: number;
+  title: string;
+  content: string;
+  anonymous: boolean;
+  images: CreateArticleImagePayload[];
+}
+
+export interface CreateArticleBody {
+  title: string;
+  content: string;
+  anonymity: boolean;
+  images: CreateArticleImagePayload[];
+}
+
+export type CreateArticleApiData = ApiSuccessResponse<{ postId: number }>;
+
+export const createArticle = (params: CreateArticleParams) => {
+  const { categoryId, anonymous, ...restParams } = params;
+  const endpoint = endpoints.articles.create(categoryId);
+
+  const body: CreateArticleBody = {
+    ...restParams,
+    anonymity: anonymous,
+  };
+
+  return privateAxios
+    .post<CreateArticleApiData>(endpoint, body)
+    .then((res) => res.data.data.postId);
+};
